Create corner.py
This code will ask user to input an image using a Tkinter GUI, after getting the input the code will aske the weight value Kh (which is mostly in between 0.04 to 0.07) and a threshold value which will decide the extent of corner detecion in image. a progress bar will keep track of the amount of progress this code is doing.